+++ src/comps/Features/Heartlist/HeartlistWrap.svelte
@@ -4,6 +4,7 @@
   import { localStore, lsKeyHeart, storedHearts, storedHeartsShare } from '@stores';
   import Heartlist from './Heartlist.svelte';
   import HeartlistShare from './HeartlistShare.svelte';
+  import Checkbox from '../../Filter/Checkbox.svelte';
 
   const flipDurationMs = 300;
   let heartLists: any;
@@ -41,10 +42,7 @@
 </div>
 
 <nav class="wrap small-margin no-h-margin">
-  <label class="checkbox">
-    <input type="checkbox" bind:checked={changeOrder} />
-    <span>Reihenfolge ändern aktivieren</span>
-  </label>
+  <Checkbox label="Reihenfolge ändern aktivieren" bind:checked={changeOrder} />
 </nav>
 
 <HeartlistShare />