+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * [20.07.2022]
   * add 12 months+ pipeline for older announcement products
+  * refactore checkbox filter to one component
 
 * [18.07.2022]
   * fix tooltip hide below producs